backend/requirements.txt
Django==3.2.3
django-filter==23.5
djangorestframework==3.12.4
django-cors-headers==3.13.0
djoser==2.1.0
drf-extra-fields==3.7.0
gunicorn==20.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
PyYAML==6.0
python-dotenv